latest changes with floats & friends

+++ gp_surrogate.py
@@ -49,7 +49,7 @@ def gp_surrogate_model(data_x, low_dim_x, test_data_x, test_data_y, seed, share_
 
     print("duplicate front length: " + str(len(front)) + " , non-duplicate front length: " + str(len(front_non_duplicate)))
     for individual in front_non_duplicate:
-        output = individual.GetOutput(test_data_x)
+        output = individual.GetOutput(test_data_x.astype(float))
         individual_output = np.empty(output.shape)
         for i in range(individual.num_sup_functions):
